Move process spawning logic to a separate class
